+++ src/util.ts
@@ -74,20 +74,3 @@ export async function getUniqueFileName(folderPath: string, fileName: string, fi
 
 	throw new Error('Could not find unique name for new file.');
 }
-
-export function removeDuplicatesById<T extends { id: string }>(entries: T[]): T[] {
-	var mapById = new Map<string, T>();
-	entries.forEach(n => {
-		mapById.set(n.id, n);
-	});
-
-	return [...mapById.values()];
-}
-
-export function truncateWithEllipses(s: string, maxCharacters) {
-	if (s && s.length > maxCharacters) {
-		return `${s.slice(0, maxCharacters)}...`;
-	}
-
-	return s;
-}
